Buffer pool (#2)
* Implement Buffer pool

* Change uint32 to int64

* Update tests and comments for Buffer pool

+++ file/block.go
@@ -4,10 +4,10 @@ import "fmt"
 
 type Block struct {
 	Filename string
-	Number   uint32
+	Number   int64
 }
 
-func GetBlock(fileName string, blockNum uint32) Block {
+func GetBlock(fileName string, blockNum int64) Block {
 	return Block{fileName, blockNum}
 }
